Simplify the Shipper base class type.
The ProtocolShipperType was causing some confusion in our static type
checking (mypy). Removing it and just using ProtocolShipper as the return
type of the shipper factory simplifies this. Explicitly subclassing
the ABC ('ProtocolShipper'), also guarantees a runtime check that
the shipper class implements the required ABC methods.

The only downside of this approach is that we cannot write this code:

```
elasticsearch_shipper: ElasticsearchShipper = create_from_output('elasticsearch', ...)
```

This seems totally reasonable - sure, the factory returns different types
depending on its inputs, but I think its better for callers not to assume
which concrete type, and just use the ABC methods only.