Send telemetry data on InputSelectedEvent
After working on the Indexer, I realized the current payload might
contain redundant data.

Instead of sending the list of configured inputs, a better option is
to send only the input sourcing the data for each lambda execution.

For example:

```JSON
{
  "function_id": "80036aa038",
  "function_version": "v0.0.0",
  "execution_id": "6b76b62589",
  "cloud_provider": "aws",
  "cloud_region": "us-east-1",
  "memory_limit_in_mb": "512",
  "start_time": "1676016439.435081",
  "input": {
    "type": "s3-sqs",
    "outputs": [
      "elasticsearch"
    ]
  }
}
```

Now the payload is focused on the lambda execution, but we can still
get the idea of which inputs each deployment uses by putting together
data from different lambda executions.