Remove call to EC2:DescribeRegions API (#811)
Remove the call to the EC2:DescribeRegions API.

We aim to increase performance and avoid throttling.

As of today, the Lambda runtime can only trigger the function from a (CloudWatch) log group in the same region as the function. 

Since the log group and function region are the same, we can use the lambda function region to look up [^1] the input in the configuration and avoid matching with all the existing regions.