[s3-sqs] Log a Lambda event summary on exception (#860)
ESF summarizes the lambda event and adds the summary as extra information to the exception log message if there is an exception when processing a lambda event from the `s3-sqs` input, 

The summary contains the essential elements for the lambda event. The `s3-sqs` input includes the bucket ARN and the object key. The summary limits the number of records to `max_records`.

ESF currently supports lambda event summary for `s3-sqs` only. We want to hear users' feedback before making additional changes or extending the summary to other input types.