Release v1.20.1 (#904)
Update the ESF version to trigger the release of v1.20.1. Also, update examples in docs and the CHANGELOG file.

This release only contains docs updates. The most important is a fix for a 404 on the `README-AWS.md` file, used as the SAR (Serverless Application Repository) content on the AWS console.